// apps/host/src/components/layout/types.ts
type BaseNavItem = {
  title: string
  badge?: string
  icon?: React.ElementType
}

type NavLink = BaseNavItem & {
  url: string
  items?: never
}

type NavCollapsible = BaseNavItem & {
  items: (BaseNavItem & { url: string })[]
  url?: never
}

type NavItem = NavCollapsible | NavLink

type NavGroup = {
  title: string
  items: NavItem[]
}

type SidebarUser = {
  name: string
  email: string
  avatar: string
}

type SidebarTeam = {
  name: string
  logo: React.ElementType
  plan: string
}

type SidebarData = {
  user: SidebarUser
  teams: SidebarTeam[]
  navGroups: NavGroup[]
}

export type { SidebarData, NavGroup, NavItem, NavCollapsible, NavLink, SidebarUser, SidebarTeam }